member-project/src/test/java/com/example/memberproject/repository/MemoryMemberRepositoryTest.java
package com.example.memberproject.repository;

import com.example.memberproject.domain.Member;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Test;

import java.util.List;

import static org.assertj.core.api.Assertions.*;

// 굳이 public으로 안해도 된다.
class MemoryMemberRepositoryTest {

    MemoryMemberRepository repository = new MemoryMemberRepository();

    @AfterEach // callback 메소드
    public void afterEach() { // test가 끝날때마다 repository를 지워줌
        repository.clearStore();
    }

    @Test
    public void save() {
        Member member = new Member();
        member.setName("spring");

        repository.save(member);

        // return type이 optional이라 그냥 get()으로 꺼낼 수 있음
        // get으로 그냥 꺼내는 것이 좋은 방법은 아니지만 test이기 때문에 그냥
        Member result = repository.findById(member.getId()).get();
        // Assertions.assertEquals(member, null); // 기대 , 결과
        assertThat(member).isEqualTo(result);
    }

    @Test
    public void findByName() {
        Member member1 = new Member();
        member1.setName("spring1");
        repository.save(member1);

        Member member2 = new Member();
        member2.setName("spring2");
        repository.save(member2);

        Member result = repository.findByName("spring1").get();

        assertThat(result).isEqualTo(member1);
    }

    @Test
    public void findAll() {
        Member member1 = new Member();
        member1.setName("spring1");
        repository.save(member1);

        Member member2 = new Member();
        member2.setName("spring2");
        repository.save(member2);

        List<Member> result = repository.findAll();

        assertThat(result.size()).isEqualTo(2);
    }
}
